pyproject.toml
[tool.poetry]
name = "codex"
version = "0.1.0"
description = "Computational Oracle for Data Encoding and eXploratio"
authors = ["SwiftyOS <craigswift13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.6.0"
pgvector = "^0.2.4"
matplotlib = "^3.8.2"
fastapi = "^0.109.1"
google-cloud-storage = "^2.14.0"
uvicorn = "^0.25.0"
click = "^8.1.7"
colorama = "^0.4.6"
black = "^23.12.0"
isort = "^5.13.2"
coloredlogs = "^15.0.1"
tenacity = "^8.2.3"
psycopg2-binary = "^2.9.9"
prisma = "^0.13.1"
python-json-logger = "^2.0.7"
pydantic = { version = "^2.6.1", extras = ["email"] }
python-multipart = "^0.0.7"
fuzzywuzzy = "^0.18.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
pytest-asyncio = "^0.23.5"
sentry-sdk = {extras = ["fast-api"], version = "^1.40.6"}
pyright = "^1.1.354"
packaging = ">=24.0"
gitpython = "^3.1.42"
tiktoken = "^0.7.0"
google-cloud = "^0.34.0"
google-auth = "^2.29.0"
pygithub = "^2.3.0"
google-api-python-client = "^2.126.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
nicegui = "^1.4.22"
langsmith = "^0.1.52"
python-levenshtein = "^0.25.1"
streamlit = "^1.35.0"
anthropic = "^0.29.0"
groq = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.6.0"
ruff = "^0.1.11"
pytest-cov = "^4.1.0"
pytest-integration = "^0.2.3"
pandas = "^2.2.1"
numpy = "^1.26.4"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
